refactor: rewrite cjsToEsm.cjs to esm.mjs
